+++ Changelog.md
@@ -32,6 +32,8 @@ Unstable features
     ([#733])
   * Fixed a panic in `QnameRouter` if a request that doesn’t have at least
     one valid name is dispatched to it. ([#734])
+* `unstable-sign`:
+  * Fixed a panic when an NSEC owner name is too long. ([#735])
 * `unstable-validator`:
   * Fixed a panic when encountering an invalid NSEC3 label. ([#728])
 * `unstable-xfr`:
@@ -62,6 +64,7 @@ Other changes
 [#732]: https://github.com/NLnetLabs/domain/pull/732
 [#733]: https://github.com/NLnetLabs/domain/pull/733
 [#734]: https://github.com/NLnetLabs/domain/pull/734
+[#735]: https://github.com/NLnetLabs/domain/pull/735
 [@SebastiaanYN]: https://github.com/SebastiaanYN